reorganized a few policies to accommodate pundit namespace hack
